+++ packages/dialtone-vue3/components/emoji_picker/emoji_picker.stories.js
@@ -52,10 +52,31 @@ const recentlyUsedEmojis = [
   },
 ];
 
+const customEmojis = [
+  {
+    name: 'shipit',
+    date_added: 1730918816847,
+    added_by: 'Ignacio Ropolo',
+    image: 'https://github.githubassets.com/images/icons/emoji/shipit.png',
+    unicode_character: '1f44d',
+  },
+  {
+    name: 'thumbs up',
+    category: 'people',
+    shortname: ':thumbsup:',
+    shortname_alternates: [':+1:', ':thumbup:'],
+    keywords: ['+1', 'hand', 'thumb', 'up', 'uc6'],
+    unicode_output: '1f44d',
+    unicode_character: '1f44d',
+  },
+];
+
 export const argsData = {
   onSkinTone: action('skin-tone'),
+  onAddEmoji: action('add-emoji'),
   onSelectedEmoji: action('selected-emoji'),
   onClose: action('close'),
+  onScrollBottomReached: action('scroll-bottom-reached'),
   tabSetLabels: [
     'Most recently used',
     'Smileys and people',
@@ -66,8 +87,11 @@ export const argsData = {
     'Objects',
     'Symbols',
     'Flags',
+    'Custom',
   ],
   recentlyUsedEmojis,
+  customEmojis,
+  addEmojiLabel: 'Add emoji',
   skinSelectorButtonTooltipLabel: 'Change default skin tone',
   searchNoResultsLabel: 'No results',
   searchResultsLabel: 'Search results',
@@ -97,7 +121,17 @@ export const argTypesData = {
       disable: true,
     },
   },
+  onAddEmoji: {
+    table: {
+      disable: true,
+    },
+  },
   onClose: {
+    table: {
+      disable: true,
+    },
+  },
+  onScrollBottomReached: {
     table: {
       disable: true,
     },