+++ packages/dialtone-vue3/components/emoji_picker/emoji_picker_default.story.vue
@@ -4,6 +4,8 @@
     :skin-selector-button-tooltip-label="$attrs.skinSelectorButtonTooltipLabel"
     :tab-set-labels="$attrs.tabSetLabels"
     :recently-used-emojis="$attrs.recentlyUsedEmojis"
+    :custom-emojis="$attrs.customEmojis"
+    :add-emoji-label="$attrs.addEmojiLabel"
     :search-results-label="$attrs.searchResultsLabel"
     :search-no-results-label="$attrs.searchNoResultsLabel"
     :search-placeholder-label="$attrs.searchPlaceholderLabel"
@@ -11,6 +13,8 @@
     :show-search="$attrs.showSearch"
     @skin-tone="skinTone = $event"
     @selected-emoji="$attrs.selectedEmoji"
+    @scroll-bottom-reached="$attrs.scrollBottomReached"
+    @add-emoji="$attrs.onAddEmoji($event)"
   />
 </template>