+++ packages/dialtone-vue3/components/rich_text_editor/extensions/mentions/MentionSuggestion.vue
@@ -1,18 +1,52 @@
 <template>
   <dt-stack
     direction="row"
+    class="d-mention-suggestion__container"
     gap="400"
   >
     <dt-avatar
       :full-name="name"
       :image-src="avatarSrc"
       :image-alt="name"
-      size="xs"
+      :show-presence="showDetails"
+      :presence="presence"
+      size="sm"
     />
-    <!-- eslint-disable-next-line vue/no-restricted-class -->
-    <span class="d-mention-suggestion__name">
-      {{ name }}
-    </span>
+    <dt-stack
+      class="d-mention-suggestion__details-container"
+      gap="100"
+    >
+      <!-- eslint-disable-next-line vue/no-restricted-class -->
+      <span class="d-mention-suggestion__name">
+        {{ name }}
+      </span>
+      <dt-stack
+        v-if="showDetails"
+        direction="row"
+        gap="300"
+        class="d-label--sm-plain"
+      >
+        <span
+          v-if="presenceText"
+          class="d-mention-suggestion__presence"
+          :class="[presenceFontColorClass]"
+        >
+          {{ presenceText }}
+        </span>
+        <div
+          v-if="status && presenceText"
+          class="d-mention-suggestion__divider"
+        >
+          •
+        </div>
+        <div
+          v-if="status"
+          class="d-mention-suggestion__status"
+        >
+          {{ status }}
+        </div>
+      </dt-stack>
+    </dt-stack>
   </dt-stack>
 </template>
 
@@ -42,6 +76,33 @@ export default {
 
     avatarSrc () {
       return this.item.avatarSrc;
+    },
+
+    presence () {
+      return this.item.presence;
+    },
+
+    status () {
+      return this.item.status;
+    },
+
+    presenceText () {
+      return this.item.presenceText;
+    },
+
+    presenceFontColorClass () {
+      const presenceFontColors = {
+        active: 'd-recipe-contact-row--active',
+        busy: 'd-recipe-contact-row--busy',
+        away: 'd-recipe-contact-row--away',
+        offline: 'd-recipe-contact-row--busy',
+      };
+
+      return presenceFontColors[this.presence];
+    },
+
+    showDetails () {
+      return this.item.showDetails;
     },
   },
 };