+++ packages/dialtone-vue3/components/rich_text_editor/mention_suggestion.js
@@ -1,54 +1,76 @@
+import defaultImage from '@/common/assets/avatar2.png';
+
 /* eslint-disable max-len */
 const CONTACT_LIST = [
   {
     id: 'test.person',
     name: 'Test Person',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
+    showDetails: true,
+    presence: 'active',
+    presenceText: '',
+    status: 'This is a test status',
   },
   {
     id: 'test.person2',
     name: 'Test Person 2',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
+    showDetails: true,
+    presence: 'busy',
+    presenceText: 'DND',
+    status: 'This is a test status',
   },
   {
     id: 'test.person3',
     name: 'Test Person 3',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
+    showDetails: true,
+    presence: 'busy',
+    presenceText: 'DND',
+    status: '',
   },
   {
     id: 'brad.paugh',
     name: 'Brad Paugh',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
+    showDetails: true,
+    presence: 'offline',
+    status: 'This would be a long status that would be truncated. This would be a long status that would be truncated. ',
   },
   {
     id: 'bradley.hawkins',
     name: 'Bradley Hawkins',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
+    showDetails: true,
+    presence: 'away',
+    presenceText: 'In a meeting',
+    status: 'This would be a long status that would be truncated. This would be a long status that would be truncated. ',
+
   },
   {
     id: 'julio.ortega',
     name: 'Tico Ortega',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
   },
   {
-    id: 'ignacio.Ropolo',
+    id: 'ignacio.ropolo',
     name: 'Ignacio Ropolo',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
   },
   {
     id: 'nina.repetto',
     name: 'Nina Repetto',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
   },
   {
     id: 'long.name',
     name: 'LongnameLongnameLongnameLongnameLongnameLongnameLongnameLongnameLongnameLongnameLongnameLongnameLongnameLongnameLongname',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
   },
   {
     id: 'long.name.with.spaces',
     name: 'Long Name Long Name Long Name Long NameLong Name Long Name Long Name Long NameLong Name Long Name Long Name Long NameLong Name Long Name Long Name Long NameLong Name Long Name Long Name Long Name ',
-    avatarSrc: 'https://avatars.githubusercontent.com/u/13851061?s=460&u=1f1b5b0b5b2b2b2b2b2b2b2b2b2b2b2b2b2b2b&v=4',
+    avatarSrc: defaultImage,
   },
 ];