+++ packages/dialtone-vue3/recipes/conversation_view/message_input/message_input.stories.js
@@ -212,6 +212,7 @@ export const argsData = {
       'Objects',
       'Symbols',
       'Flags',
+      'Custom',
     ],
     skinTone: 'Default',
   },